rboot: 76 tests de unidad y seis fallos del cargador

+++ rboot/src/idt.rs
@@ -9,7 +9,7 @@ use core::sync::atomic::{AtomicU64, AtomicUsize, Ordering};
 use uefi::proto::console::gop::ModeInfo;
 use x86_64::structures::idt::{InterruptDescriptorTable, InterruptStackFrame, PageFaultErrorCode};
 
-use crate::progress;
+use rboot::progress;
 
 static FB_ADDR: AtomicU64 = AtomicU64::new(0);
 static STRIDE: AtomicUsize = AtomicUsize::new(0);
@@ -21,9 +21,9 @@ static mut IDT: InterruptDescriptorTable = InterruptDescriptorTable::new();
 pub fn init(mode: ModeInfo, fb_addr: u64) {
     let (sw, sh) = mode.resolution();
     FB_ADDR.store(fb_addr, Ordering::SeqCst);
-    STRIDE.store(mode.stride() as usize, Ordering::SeqCst);
-    SW.store(sw as usize, Ordering::SeqCst);
-    SH.store(sh as usize, Ordering::SeqCst);
+    STRIDE.store(mode.stride(), Ordering::SeqCst);
+    SW.store(sw, Ordering::SeqCst);
+    SH.store(sh, Ordering::SeqCst);
 
     unsafe {
         let idt = &mut *core::ptr::addr_of_mut!(IDT);
